Move model into parser for now
